requirements.txt
aiohttp==3.9.1
websockets==12.0
pytest==7.4.3
python-dotenv==1.0.0
asyncio==3.4.3
requests==2.31.0
logging==0.4.9.6 
supabase==2.11.0
